# boards/stm32f769i-disco: enable FMC with SDRAM support