--- a/boards/stm32f769i-disco/Kconfig
+++ b/boards/stm32f769i-disco/Kconfig
@@ -14,6 +14,9 @@ config BOARD_STM32F769I_DISCO
     select CPU_MODEL_STM32F769NI
 
     # Put defined MCU peripherals here (in alphabetical order)
+    select HAS_PERIPH_FMC
+    select HAS_PERIPH_FMC_SDRAM
+    select HAS_PERIPH_FMC_32BIT
     select HAS_PERIPH_RTC
     select HAS_PERIPH_RTT
     select HAS_PERIPH_TIMER