--- a/boards/stm32f769i-disco/Makefile.features
+++ b/boards/stm32f769i-disco/Makefile.features
@@ -2,6 +2,9 @@ CPU = stm32
 CPU_MODEL = stm32f769ni
 
 # Put defined MCU peripherals here (in alphabetical order)
+FEATURES_PROVIDED += periph_fmc
+FEATURES_PROVIDED += periph_fmc_32bit
+FEATURES_PROVIDED += periph_fmc_sdram
 FEATURES_PROVIDED += periph_rtc
 FEATURES_PROVIDED += periph_rtt
 FEATURES_PROVIDED += periph_timer