@gschorcht 2023-07-28
@dylad According to the authorship of the board definition, I suspect you have a `stm32f769i-disco` board, which I do not. May I kindly ask you to test the PR on this board? Testing FMC on this board would be very interesting since it is the only supported board so far that uses FMC and SDRAM with 32-bit data bus width. FMC and SDRAM with 16-bit data bus width was already tested extensively in PR #19843.

The PR itself is not that spectacular, we have already used the configuration provided in this PR several times in PR #19843 and it is simply copied from such a board. The only difference are the additional 16 data signals.