@dylad 2023-07-28
@gschorcht Sorry I don't have this one, I just had a look at my ST board.
Best I can do is nucleo-F746ZG but I don't know if it will help you.