@gschorcht 2023-07-28
> Best I can do is nucleo-F746ZG but I don't know if it will help you.

Hm, too bad. Nucleo boards do not have external RAM. I have a "stm32f746g-disco" board on which I could test the FMC peripheral with SDRAM and 16-bit data bus.