@gschorcht 2023-08-23
I ordered a `stmf769i-disco` board and tested the FMC with 32-bit data bus width. It works as expected see https://github.com/RIOT-OS/RIOT/pull/19851#issue-1826367528.